fix : np.int depreciated use np.int_

+++ lightning-hydra-template/src/models/resnet50_cutmix_module.py
@@ -98,8 +98,8 @@ class Resnet50_cutmixModule(LightningModule):
         W = size[2]
         H = size[3]
         cut_rat = np.sqrt(1. - lam)
-        cut_w = np.int(W * cut_rat)
-        cut_h = np.int(H * cut_rat)
+        cut_w = np.int_(W * cut_rat)
+        cut_h = np.int_(H * cut_rat)
 
         # uniform
         cx = np.random.randint(W)
